+++ packages/eslint-plugin/src/rules/prefer-literal-enum-member.ts
@@ -6,8 +6,7 @@ export default createRule({
   meta: {
     type: 'suggestion',
     docs: {
-      description:
-        'Require that all enum members be literal values to prevent unintended enum member name shadow issues',
+      description: 'Require all enum members to be literal values',
       recommended: 'strict',
       requiresTypeChecking: false,
     },