+++ packages/website/src/components/editor/createProvideCodeActions.ts
@@ -1,6 +1,9 @@
 import type Monaco from 'monaco-editor';
-import type { LintCodeAction } from '../linter/lintCode';
-import { createURI } from '../linter/utils';
+import {
+  createEditOperation,
+  createURI,
+  LintCodeAction,
+} from '../linter/utils';
 
 export function createProvideCodeActions(
   fixes: Map<string, LintCodeAction[]>,
@@ -24,8 +27,6 @@ export function createProvideCodeActions(
       for (const marker of context.markers) {
         const messages = fixes.get(createURI(marker)) ?? [];
         for (const message of messages) {
-          const start = model.getPositionAt(message.fix.range[0]);
-          const end = model.getPositionAt(message.fix.range[1]);
           actions.push({
             title: message.message,
             diagnostics: [marker],
@@ -34,15 +35,7 @@ export function createProvideCodeActions(
               edits: [
                 {
                   resource: model.uri,
-                  edit: {
-                    range: {
-                      startLineNumber: start.lineNumber,
-                      startColumn: start.column,
-                      endLineNumber: end.lineNumber,
-                      endColumn: end.column,
-                    },
-                    text: message.fix.text,
-                  },
+                  edit: createEditOperation(model, message),
                 },
               ],
             },